pyproject.toml
[tool.poetry]
name = "planetmint-transactions"
version = "0.7.0"
description = "Python implementation of the planetmint transactions spec"
authors = ["Lorenz Herzberger <lorenzherzberger@gmail.com>"]
readme = "README.md"
packages = [
    { include = "transactions" }
]

[tool.poetry.dependencies]
python = "^3.9"
planetmint-cryptoconditions = "^1.1.0"
base58 = "^2.1.1"
planetmint-py-cid = "^0.4.2"
planetmint-ipld = "^0.0.3"
python-rapidjson = "^1.8"
jsonschema = "^4.16.0"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "23.1.0"
hypothesis = "^6.54.6"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

